Add counted versions of Thrust partition functions
Add counted version of each Thrust partition function.  These new
functions return the number of items in the first section of the
partition, rather than an input iterator.

This functionality can be convenient in general, but it was actually
necessary to work around a Cython compilation issue, where a subtraction
between two zip iterators could not be compiled.  This meant that it was
not possible to determine the split point after partitioning.