Add DeviceVectorCollection...
Add `DeviceVectorCollection` to store collections of `DeviceVector`
instances, where a `DeviceVectorView` is also created for each
`DeviceVector`, enabling easy slicing of vectors for Thrust functions.

When possible, it is preferable to use the `DeviceDataFrame`, but the
`DeviceDataFrame` requires that all input arrays are of the same length.
The `DeviceVectorCollection` may contain arrays of varying length.