Add `cythrust::reduce_n` function
Add `cythrust::reduce_n` function to reduce based on an input iterator and a
count rather than an input begin and end iterator (similar to
`thrust::copy_n`). This is helpful, for example, in situations where the input
iterator is constructed on-the-fly as a complex zip iterator type since only
the begin iterator must be specified.

The `cythrust::reduce_n` is implemented as a (thin) wrapper around
`thrust::reduce`.